# cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.GenerateLoad:
    properties:
      encryptionMode:
        enum:
        - rsa
        - ed25519
        - ecdsa
        - dsa
        type: string
    required:
    - encryptionMode
    type: object
  dto.GenerateSSH:
    properties:
      encryptionMode:
        enum:
        - rsa
        - ed25519
        - ecdsa
        - dsa
        type: string
      password:
        type: string
    required:
    - encryptionMode
    type: object
...
  dto.Operate:
    properties:
      operation:
        type: string
    required:
    - operation
    type: object
...
  dto.SSHConf:
    properties:
      file:
        type: string
    type: object
  dto.SSHHistory:
    properties:
      address:
        type: string
      area:
        type: string
      authMode:
        type: string
      date:
        type: string
      dateStr:
        type: string
      message:
        type: string
      port:
        type: string
      status:
        type: string
      user:
        type: string
    type: object
  dto.SSHInfo:
    properties:
      listenAddress:
        type: string
      message:
        type: string
      passwordAuthentication:
        type: string
      permitRootLogin:
        type: string
      port:
        type: string
      pubkeyAuthentication:
        type: string
      status:
        type: string
      useDNS:
        type: string
    type: object
  dto.SSHLog:
    properties:
      failedCount:
        type: integer
      logs:
        items:
          $ref: '#/definitions/dto.SSHHistory'
        type: array
      successfulCount:
        type: integer
      totalCount:
        type: integer
    type: object
  dto.SSHLogAnalysis:
    properties:
      address:
        type: string
      area:
        type: string
      failedCount:
        type: integer
      status:
        type: string
      successfulCount:
        type: integer
    type: object
...
  dto.SearchForAnalysis:
    properties:
      orderBy:
        enum:
        - Success
        - Failed
        type: string
    required:
    - orderBy
    type: object
...
  dto.SearchSSHLog:
    properties:
      Status:
        enum:
        - Success
        - Failed
        - All
        type: string
      info:
        type: string
      page:
        type: integer
      pageSize:
        type: integer
    required:
    - Status
    - page
    - pageSize
    type: object
...
  dto.SettingUpdate:
    properties:
      key:
        type: string
      value:
        type: string
    required:
    - key
    type: object
...
  dto.Upgrade:
    properties:
      version:
        type: string
    type: object
...
host: localhost
info:
  contact: {}
  description: 开源Linux面板
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 1Panel
  version: "1.0"
paths:
...
  /host/conffile/update:
    post:
      consumes:
      - application/json
      description: 上传文件更新 SSH 配置
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SSHConf'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update host SSH setting by file
      tags:
      - SSH
      x-panel-log:
        BeforeFunctions: []
        bodyKeys: []
        formatEN: update SSH conf
        formatZH: 修改 SSH 配置文件
        paramKeys: []
  /host/ssh/conf:
    get:
      description: 获取 SSH 配置文件
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Load host SSH conf
      tags:
      - SSH
  /host/ssh/generate:
    post:
      consumes:
      - application/json
      description: 生成 SSH 密钥
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GenerateSSH'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Generate host SSH secret
      tags:
      - SSH
      x-panel-log:
        BeforeFunctions: []
        bodyKeys: []
        formatEN: generate SSH secret
        formatZH: '生成 SSH 密钥 '
        paramKeys: []
  /host/ssh/log:
    post:
      consumes:
      - application/json
      description: 获取 SSH 登录日志
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SearchSSHLog'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SSHLog'
      security:
      - ApiKeyAuth: []
      summary: Load host SSH logs
      tags:
      - SSH
  /host/ssh/log/analysis:
    post:
      consumes:
      - application/json
      description: 分析 SSH 登录日志
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SearchForAnalysis'
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.SSHLogAnalysis'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Analysis host SSH logs
      tags:
      - SSH
  /host/ssh/operate:
    post:
      consumes:
      - application/json
      description: 修改 SSH 服务状态
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.Operate'
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Operate SSH
      tags:
      - SSH
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - operation
        formatEN: '[operation] SSH'
        formatZH: '[operation] SSH '
        paramKeys: []
  /host/ssh/search:
    post:
      description: 加载 SSH 配置信息
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SSHInfo'
      security:
      - ApiKeyAuth: []
      summary: Load host SSH setting info
      tags:
      - SSH
  /host/ssh/secret:
    post:
      consumes:
      - application/json
      description: 获取 SSH 密钥
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GenerateLoad'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Load host SSH secret
      tags:
      - SSH
  /host/ssh/update:
    post:
      consumes:
      - application/json
      description: 更新 SSH 配置
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SettingUpdate'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update host SSH setting
      tags:
      - SSH
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - key
        - value
        formatEN: update SSH setting [key] => [value]
        formatZH: 修改 SSH 配置 [key] => [value]
        paramKeys: []
...
  /settings/upgrade:
    get:
      consumes:
      - application/json
      description: 获取版本 release notes
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.Upgrade'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Load release notes by version
      tags:
      - System Setting
    post:
      consumes:
      - application/json
      description: 系统更新
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.Upgrade'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Upgrade
      tags:
      - System Setting
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - version
        formatEN: upgrade service => [version]
        formatZH: 更新系统 => [version]
        paramKeys: []
...
swagger: "2.0"